moltbot/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moltbot"
version = "0.1.0"
description = "Bot de infraestructura — procesa comandos y facturas via RabbitMQ"
requires-python = ">=3.11"
dependencies = [
    "pika",
    "psycopg2-binary",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff",
]

[project.scripts]
moltbot = "moltbot.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/moltbot"]

[tool.ruff]
target-version = "py311"
line-length = 100
